[Plugin] Option to export lape-style wrappers
Added the option to use lape-style export wrapper-functions to plugins.
By avoiding using the FFI, functions ...
- avoids leaking the result, if the result is an array.
- has around 5 times less call overhead

To export lape-style wrapper functions an ekstra kw have to be added at
the end of each function-string: `external;`.

EG:
>> procedure SomeFunc(const Params: PParamArray; const Result: Pointer);
cdecl; export;
...
>> ProcAddr := @SomeFunc;
>> StrPCopy(ProcDef, 'function SomeFunc(Size:Int32): TIntegerArray;
external;');
That is how we know that we should not use the FFI.

Former-commit-id: 2e6e7e01b6769aa0b75bfcbf2dbc6124d7f8de89